[AC-991] removing the OldChangePlanForm and related code
