+++ cypress/fixtures/account/200.get.starter-plan.json
@@ -40,7 +40,6 @@
           "isGuideAtBottom": true
         },
         "recipientValidationV2": false,
-        "account_feature_limits": true,
         "recipient_validation": false,
         "feature_signals_v2": true,
         "Send Test Email": null,
