+++ src/actions/billing.js
@@ -1,7 +1,6 @@
 import { formatContactData } from 'src/helpers/billing';
 import { fetch as fetchAccount, getBillingInfo } from './account';
 import { list as getSendingIps } from './sendingIps';
-import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
 import sparkpostApiRequest from 'src/actions/helpers/sparkpostApiRequest';
 import zuoraRequest from 'src/actions/helpers/zuoraRequest';
 
@@ -14,38 +13,22 @@ export function updateSubscription({ code, bundle = code, promoCode, meta = {} }
   const fetchAccountAction = () =>
     fetchAccount({ include: 'usage', meta: { onSuccess: getBillingAction } });
 
-  return (dispatch, getState) =>
-    isAccountUiOptionSet('account_feature_limits')(getState()) //TODO: Remove this + first action
-      ? dispatch(
-          sparkpostApiRequest({
-            type: 'UPDATE_SUBSCRIPTION',
-            meta: {
-              method: 'PUT',
-              url: '/v1/billing/subscription/bundle',
-              data: {
-                promo_code: promoCode,
-                bundle,
-              },
-              ...meta,
-              onSuccess: meta.onSuccess ? meta.onSuccess : fetchAccountAction,
-            },
-          }),
-        )
-      : dispatch(
-          sparkpostApiRequest({
-            type: 'UPDATE_SUBSCRIPTION',
-            meta: {
-              method: 'PUT',
-              url: '/v1/account/subscription',
-              data: {
-                promo_code: promoCode,
-                code,
-              },
-              ...meta,
-              onSuccess: meta.onSuccess ? meta.onSuccess : fetchAccountAction,
-            },
-          }),
-        );
+  return dispatch =>
+    dispatch(
+      sparkpostApiRequest({
+        type: 'UPDATE_SUBSCRIPTION',
+        meta: {
+          method: 'PUT',
+          url: '/v1/billing/subscription/bundle',
+          data: {
+            promo_code: promoCode,
+            bundle: bundle,
+          },
+          ...meta,
+          onSuccess: meta.onSuccess ? meta.onSuccess : fetchAccountAction,
+        },
+      }),
+    );
 }
 
 export function syncSubscription({ meta = {} } = {}) {
