+++ src/pages/billing/ChangePlanPage.container.js
@@ -4,17 +4,13 @@ import { ChangePlanProvider } from './context/ChangePlanContext';
 import { getBillingInfo } from 'src/actions/account';
 import { getSubscription, getBundles, getPlans, getBillingCountries } from 'src/actions/billing';
 import { selectAvailableBundles, selectAccountBilling } from 'src/selectors/accountBillingInfo';
-import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
 import { connect } from 'react-redux';
 
-const ChangePlanPageContainer = ({ newChangePlan, ...props }) =>
-  newChangePlan ? ( //TODO: Remove conditional check after replacing changeplanform
-    <ChangePlanProvider {...props}>
-      <ChangePlanPage />
-    </ChangePlanProvider>
-  ) : (
+const ChangePlanPageContainer = ({ newChangePlan, ...props }) => (
+  <ChangePlanProvider {...props}>
     <ChangePlanPage />
-  );
+  </ChangePlanProvider>
+);
 
 const mapDispatchToProps = {
   getSubscription,
@@ -34,7 +30,6 @@ const mapStateToProps = state => {
     loading: countriesLoading || plansLoading || bundlesLoading || accountLoading,
     billingCountries: state.billing.countries,
     account,
-    newChangePlan: isAccountUiOptionSet('account_feature_limits')(state),
   };
 };
