+++ src/pages/billing/ChangePlanPage.js
@@ -3,16 +3,11 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { Page } from '@sparkpost/matchbox';
 import { PendingPlanBanner } from './components/Banners';
-import OldChangePlanForm from './forms/ChangePlanForm'; //TODO: Replace in AC-991
 import { canChangePlanSelector } from 'src/selectors/accountBillingInfo';
-import NewChangePlanForm from './forms/NewChangePlanForm';
-import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
+import ChangePlanForm from './forms/ChangePlanForm';
 
 export class ChangePlanPage extends Component {
   render() {
-    //TODO: Remove in AC-991
-    const ChangePlanForm = this.props.newChangePlan ? NewChangePlanForm : OldChangePlanForm;
-
     return (
       <Page
         breadcrumbAction={{ content: 'Back to billing', to: '/account/billing', Component: Link }}
@@ -28,7 +23,6 @@ const mapStateToProps = state => ({
   account: state.account,
   subscription: state.billing.subscription || {},
   canChangePlan: canChangePlanSelector(state),
-  newChangePlan: isAccountUiOptionSet('account_feature_limits')(state),
 });
 
 export default connect(mapStateToProps)(ChangePlanPage);
