+++ src/pages/billing/forms/ChangePlanForm.js
@@ -1,192 +1,239 @@
-/* eslint max-lines: ["error", 200] */
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { reduxForm, formValueSelector } from 'redux-form';
+/* eslint max-lines: ["error", 250] */
+import React, { useEffect, useState, useCallback, useMemo } from 'react';
 import { withRouter } from 'react-router-dom';
+import { connect } from 'react-redux';
+import { reduxForm } from 'redux-form';
 import qs from 'query-string';
-import { fetch as fetchAccount, getPlans, getBillingInfo } from 'src/actions/account';
-import { updateSubscription, getBillingCountries, verifyPromoCode, clearPromoCode } from 'src/actions/billing';
+import _ from 'lodash';
+import PlanSelectSection, { SelectedPlan } from '../components/PlanSelect';
+import CurrentPlanSection from '../components/CurrentPlanSection';
+import useRouter from 'src/hooks/useRouter';
+import SubmitSection from '../components/SubmitSection';
+import CardSection from '../components/CardSection';
+import { isAws } from 'src/helpers/conditions/account';
+import { not } from 'src/helpers/conditions';
+import AccessControl from 'src/components/auth/AccessControl';
+
+import FeatureChangeSection from '../components/FeatureChangeSection';
+import { FeatureChangeContextProvider } from '../context/FeatureChangeContext';
+import { verifyPromoCode, clearPromoCode, updateSubscription } from 'src/actions/billing';
 import billingCreate from 'src/actions/billingCreate';
 import billingUpdate from 'src/actions/billingUpdate';
 import { showAlert } from 'src/actions/globalAlert';
-import { changePlanInitialValues } from 'src/selectors/accountBillingForms';
+//Selectors
 import {
-  currentPlanSelector, canUpdateBillingInfoSelector, selectTieredVisiblePlans, selectAccountBilling
+  currentPlanSelector,
+  canUpdateBillingInfoSelector,
+  getPromoCodeObject,
 } from 'src/selectors/accountBillingInfo';
-import { Panel, Grid } from '@sparkpost/matchbox';
-import { Loading, PlanPicker, ApiErrorBanner } from 'src/components';
-import Confirmation from '../components/Confirmation';
-import CardSection from '../components/CardSection';
-import promoCodeValidate from '../helpers/promoCodeValidate';
-import { isAws, isSelfServeBilling } from 'src/helpers/conditions/account';
+import { changePlanInitialValues } from 'src/selectors/accountBillingForms';
 import { selectCondition } from 'src/selectors/accessConditionState';
-import { not } from 'src/helpers/conditions';
-import * as conversions from 'src/helpers/conversionTracking';
-import AccessControl from 'src/components/auth/AccessControl';
+
+import { ApiErrorBanner } from 'src/components';
+import { Loading } from 'src/components/loading/Loading';
+import styles from './ChangePlanForm.module.scss';
+import { useChangePlanContext } from '../context/ChangePlanContext';
 import { prepareCardInfo } from 'src/helpers/billing';
-import _ from 'lodash';
 
 const FORMNAME = 'changePlan';
 
-export class ChangePlanForm extends Component {
+export const ChangePlanForm = ({
+  location,
+  //redux form
+  handleSubmit,
+  submitting,
+  //Redux Props
+  currentPlan,
+  canUpdateBillingInfo,
+  history,
+  //Redux Actions
+  verifyPromoCode,
+  promoCodeObj,
+  clearPromoCode,
+  updateSubscription,
+  billingUpdate,
+  billingCreate,
+  showAlert,
+  ...restProps
+}) => {
+  const { billingCountries, account, bundles, loading, error } = useChangePlanContext();
 
-  state = {
-    useSavedCC: null
-  };
+  const { requestParams: { code, promo } = {}, updateRoute } = useRouter();
 
-  componentDidMount() {
-    this.props.getPlans().then(() => {
-      const { initialValues, selectedPlan } = this.props;
+  const [selectedBundleCode, selectBundle] = useState(code);
+  const bundlesByCode = useMemo(() => _.keyBy(bundles, 'bundle'), [bundles]);
+  const selectedBundle = bundlesByCode[selectedBundleCode];
+  const isDowngradeToFree = _.get(selectedBundle, 'messaging.price') <= 0;
 
-      //After loading plans, verifies query promo against plan and applies UI discount
-      //TODO: Move to componentDidUpdate
-      if (initialValues.promoCode && !_.isEmpty(selectedPlan)) {
-        return this.props.verifyPromoCode({
-          promoCode: initialValues.promoCode,
-          billingId: selectedPlan.billingId,
-          meta: { promoCode: initialValues.promoCode, showErrorAlert: false }
-        });
-      }
-    });
-    this.props.getBillingCountries();
-    this.props.fetchAccount();
-    this.props.getBillingInfo();
-  }
-
-  componentWillReceiveProps(nextProps) {
-    // Null check to make sure this only runs once
-    if (nextProps.canUpdateBillingInfo && this.state.useSavedCC === null) {
-      this.setState({ useSavedCC: true });
+  const onSelect = bundle => {
+    if (!bundle) {
+      clearPromoCode();
+      updateRoute({});
     }
-  }
-
-  handleCardToggle = () => {
-    this.setState({ useSavedCC: !this.state.useSavedCC });
+    selectBundle(bundle);
   };
 
-  onSubmit = (values) => {
-    const { account, billing, updateSubscription, billingCreate, billingUpdate, showAlert, history, verifyPromoCode } = this.props;
-    const oldCode = account.subscription.code;
-    const newCode = values.planpicker.code;
-    const isDowngradeToFree = values.planpicker.isFree;
-    const selectedPromo = billing.selectedPromo;
-    const newValues = values.card && !isDowngradeToFree
-      ? { ...values, card: prepareCardInfo(values.card) }
-      : values;
+  const [useSavedCC, setUseSavedCC] = useState(null);
+  const handleCardToggle = () => setUseSavedCC(!useSavedCC);
+  useEffect(() => {
+    if (canUpdateBillingInfo && useSavedCC === null) {
+      setUseSavedCC(true);
+    }
+  }, [canUpdateBillingInfo, useSavedCC]);
+
+  const isPlanSelected = Boolean(selectedBundle && currentPlan.plan !== selectedBundle.bundle);
+  const applyPromoCode = useCallback(
+    promoCode => {
+      const billingId = _.get(selectedBundle, 'messaging.billing_id');
+      if (billingId) {
+        verifyPromoCode({ promoCode, billingId, meta: { promoCode, showErrorAlert: false } });
+      }
+    },
+    [selectedBundle, verifyPromoCode],
+  );
+
+  //Applies promo code if in query param
+  useEffect(() => {
+    if (promo && selectedBundle) {
+      applyPromoCode(promo);
+    }
+  }, [applyPromoCode, promo, selectedBundle, verifyPromoCode]);
+  const onSubmit = values => {
+    const { isAws } = restProps;
+    const newCode = selectedBundleCode;
+    const { selectedPromo } = promoCodeObj;
+    const billingId = _.get(selectedBundle, 'messaging.billing_id');
+
+    const cardValues =
+      values.card && !isDowngradeToFree
+        ? { ...values, card: prepareCardInfo(values.card) }
+        : values;
+
+    const newValues = {
+      ...cardValues,
+      bundle: newCode,
+      billingId,
+    };
     let action = Promise.resolve({});
+
     if (!_.isEmpty(selectedPromo) && !isDowngradeToFree) {
       newValues.promoCode = selectedPromo.promoCode;
-      action = verifyPromoCode({ promoCode: selectedPromo.promoCode , billingId: values.planpicker.billingId, meta: { promoCode: selectedPromo.promoCode }});
+      action = verifyPromoCode({
+        promoCode: selectedPromo.promoCode,
+        billingId,
+        meta: { promoCode: selectedPromo.promoCode },
+      });
     }
     return action
       .then(({ discount_id }) => {
         newValues.discountId = discount_id;
         // decides which action to be taken based on
         // if it's aws account, it already has billing and if you use a saved CC
-        if (this.props.isAws) {
+        if (isAws) {
           return updateSubscription({ code: newCode });
         } else if (account.billing) {
-          return this.state.useSavedCC || isDowngradeToFree ? updateSubscription({ code: newCode, promoCode: selectedPromo.promoCode }) : billingUpdate(newValues);
+          return useSavedCC || isDowngradeToFree
+            ? updateSubscription({ bundle: newCode, promoCode: selectedPromo.promoCode })
+            : billingUpdate(newValues);
         } else {
           return billingCreate(newValues); // creates Zuora account
         }
       })
+      .then(() => clearPromoCode()) //TODO: Consolidate state of promocode between redux form and hook on PromoCodeNew
       .then(() => history.push('/account/billing'))
-      .then(() => {
-        conversions.trackPlanChange({ allPlans: billing.plans, oldCode, newCode });
-        return showAlert({ type: 'success', message: 'Subscription Updated' });
-      });
+      .then(() => showAlert({ type: 'success', message: 'Subscription Updated' }));
   };
 
-  onPlanSelect = (e) => {
-    const { currentPlan, clearPromoCode } = this.props;
-    if (currentPlan.code !== e.code) {
-      clearPromoCode();
-    }
-  }
-
-  render() {
-    const { loading, canUpdateBillingInfo, submitting, currentPlan, selectedPlan, plans, isSelfServeBilling, billing, verifyPromoCode, error, account } = this.props;
-    const { countries } = billing;
-    if (error) {
-      return (
-        <ApiErrorBanner
-          status='danger'
-          message="We couldn't render the page. Reload to try again."
-          reload={() => location.reload()}
-          errorDetails={error.message}
-        />
-      );
-    }
-
-    if (loading) {
-      return <Loading />;
-    }
-
-    // Manually billed accounts can submit without changing plan
-    const disableSubmit = submitting ||
-      (isSelfServeBilling && currentPlan.code === selectedPlan.code) ||
-      // do not allow private, deprecated, etc. plans to enable billing
-      (selectedPlan.status !== 'public' && selectedPlan.status !== 'secret') ||
-      // can't submit while verifying promoCode
-      billing.promoPending;
-
+  if (error) {
     return (
-      <form onSubmit={this.props.handleSubmit(this.onSubmit)}>
-        <Grid>
-          <Grid.Column>
-            <Panel>
-              {!_.isEmpty(plans)
-                ? <PlanPicker disabled={submitting} plans={plans} onChange={this.onPlanSelect}/>
-                : null
-              }
-            </Panel>
-            <AccessControl condition={not(isAws)}>
-              <CardSection
-                account={account}
-                countries={countries}
-                selectedPlan={selectedPlan}
-                submitting={submitting}
-                canUpdateBillingInfo={canUpdateBillingInfo}
-                useSavedCC={this.state.useSavedCC}
-                handleCardToggle={this.handleCardToggle}
-              />
-            </AccessControl>
-          </Grid.Column>
-          <Grid.Column xs={12} md={5}>
-            <Confirmation
-              verifyPromoCode={verifyPromoCode}
-              selectedPromo={billing.selectedPromo}
-              current={currentPlan}
-              selected={selectedPlan}
-              billingEnabled={isSelfServeBilling}
-              disableSubmit={disableSubmit} />
-          </Grid.Column>
-        </Grid>
-      </form>
+      <ApiErrorBanner
+        status="danger"
+        message="We couldn't render the page. Reload to try again."
+        reload={() => location.reload()}
+        errorDetails={error.message}
+      />
     );
   }
-}
+
+  if (loading) {
+    return <Loading />;
+  }
+
+  return (
+    <form onSubmit={handleSubmit(onSubmit)}>
+      <div className={styles.ChangePlanForm}>
+        <div className={styles.CurrentPlanSection}>
+          <CurrentPlanSection currentPlan={currentPlan} isPlanSelected={Boolean(selectedBundle)} />
+        </div>
+        <div className={styles.MainContent}>
+          {isPlanSelected ? (
+            <>
+              <SelectedPlan
+                bundle={selectedBundle}
+                onChange={onSelect}
+                promoCodeObj={promoCodeObj}
+                handlePromoCode={{
+                  applyPromoCode: applyPromoCode,
+                  clearPromoCode: clearPromoCode,
+                }}
+              />
+              <FeatureChangeContextProvider selectedBundle={selectedBundle}>
+                <FeatureChangeSection />
+                {!isDowngradeToFree && (
+                  <AccessControl condition={not(isAws)}>
+                    <CardSection
+                      account={account}
+                      countries={billingCountries}
+                      selectedPlan={selectedBundle}
+                      canUpdateBillingInfo={canUpdateBillingInfo}
+                      submitting={submitting}
+                      isNewChangePlanForm={true} //TODO: remove this when removing the OldChangePlanForm
+                      defaultToggleState={!useSavedCC}
+                      handleCardToggle={handleCardToggle}
+                    />
+                  </AccessControl>
+                )}
+
+                <SubmitSection
+                  loading={submitting}
+                  selectedBundle={selectedBundle}
+                  account={account}
+                  brightbackCondition={isDowngradeToFree}
+                />
+              </FeatureChangeContextProvider>
+            </>
+          ) : (
+            <PlanSelectSection onSelect={onSelect} bundles={bundles} currentPlan={currentPlan} />
+          )}
+        </div>
+      </div>
+    </form>
+  );
+};
 
 const mapStateToProps = (state, props) => {
-  const selector = formValueSelector(FORMNAME);
   const { code: planCode, promo: promoCode } = qs.parse(props.location.search);
-  const plans = selectTieredVisiblePlans(state);
-  const { account, loading } = selectAccountBilling(state);
   return {
-    loading: (!account.created && loading) || (_.isEmpty(plans) && state.billing.plansLoading),
-    isAws: selectCondition(isAws)(state),
-    account,
-    billing: state.billing,
+    initialValues: changePlanInitialValues(state, { planCode, promoCode }),
     canUpdateBillingInfo: canUpdateBillingInfoSelector(state),
-    isSelfServeBilling: selectCondition(isSelfServeBilling)(state),
-    plans,
     currentPlan: currentPlanSelector(state),
-    selectedPlan: selector(state, 'planpicker') || {},
-    initialValues: changePlanInitialValues(state, { planCode, promoCode })
+    isAws: selectCondition(isAws)(state),
+    promoCodeObj: getPromoCodeObject(state),
   };
 };
 
-const mapDispatchtoProps = { getBillingInfo, billingCreate, billingUpdate, updateSubscription, showAlert, getPlans, getBillingCountries, fetchAccount, verifyPromoCode, clearPromoCode };
-const formOptions = { form: FORMNAME, enableReinitialize: true, asyncValidate: promoCodeValidate(FORMNAME), asyncChangeFields: ['planpicker'], asyncBlurFields: ['promoCode']};
-export default withRouter(connect(mapStateToProps, mapDispatchtoProps)(reduxForm(formOptions)(ChangePlanForm)));
+const formOptions = {
+  form: FORMNAME,
+  enableReinitialize: true,
+};
+
+export default withRouter(
+  connect(mapStateToProps, {
+    verifyPromoCode,
+    clearPromoCode,
+    updateSubscription,
+    billingUpdate,
+    billingCreate,
+    showAlert,
+  })(reduxForm(formOptions)(ChangePlanForm)),
+);
