+++ src/pages/billing/forms/ChangePlanForm.module.scss
@@ -14,7 +14,7 @@
 .CurrentPlanSection {
   grid-column: 1;
   grid-row: 1;
-  
+
   @media (min-width: breakpoint('large')) {
     grid-column: 2;
     grid-row: 1;
